Move enums to src/entities/enums.
