+++ packages/tests/integration/src/entities/enums/FavoriteShape.ts
