+++ packages/tests/integration/src/entities/enums/ImageType.ts
