+++ packages/tests/uuid-ids/src/entities/entities.ts
@@ -3,7 +3,7 @@
 // This file drives our import order to avoid undefined errors
 // when the subclasses extend the base classes, see:
 // https://medium.com/visual-development/how-to-fix-nasty-circular-dependency-issues-once-and-for-all-in-javascript-typescript-a04c987cf0de
-export * from "./codegen/BookStatus";
+export * from "./enums/BookStatus";
 
 export * from "./codegen/AuthorCodegen";
 export * from "./codegen/BookCodegen";
